# dist/acmeair/swagger-auth.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a AcmeAir NodeJS benchmark application Authorization Service server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  ",
        "version": "1.0.0",
        "title": "AcmeAir NodeJS Authentication Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yanglei@us.ibm.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost",
    "basePath": "/rest/api/auth-service",
    "tags": [
        {
            "name": "auth-service",
            "description": "authentication"
        },
        {
            "name": "status",
            "description": "status"
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
    	"/checkstatus":{
    		"get":{
                "tags": [
                    "status"
                ],
                "summary": "Check status",
                "description": "Check status",
                "operationId": "checkStatus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "successful operation."
                    }
                },
                "security": [
                ]
    		}
    	},
    	"/authtoken/byuserid/{userId}":{
    		"post":{
                "tags": [
                    "auth-service"
                ],
                "summary": "Create authentication token",
                "description": "Create authentication token",
                "operationId": "createToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "description": "login credential",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation.",
                        "schema": {
                        	"$ref": "#/definitions/Token"
                        }
                    },
                    "403": {
                        "description": "Invalid"
                    },
                    "404": {
                        "description": "Failure"
                    }
                },
                "security": [
                ]
            
    		}
    	},
    	"/authtoken/{tokenId}":{
    		"get":{
                "tags": [
                    "auth-service"
                ],
                "summary": "Validate authentication token",
                "description": "Validate authentication token",
                "operationId": "validateToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "tokenId",
                        "description": "token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation.",
                        "schema": {
                        	"$ref": "#/definitions/Token"
                        }
                    },
                    "403": {
                        "description": "Invalid"
                    },
                    "404": {
                        "description": "Failure"
                    }
                },
                "security": [
                ]
    		}
    	},
    	"/authtoken/{token}":{
    		"delete":{
                "tags": [
                    "auth-service"
                ],
                "summary": "Invalidate authentication token",
                "description": "Invalidate authentication token",
                "operationId": "invalidateToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "tokenId",
                        "description": "token id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation."
                    },
                    "403": {
                        "description": "Invalid"
                    },
                    "404": {
                        "description": "Failure"
                    }
                },
                "security": [
                ]
    		
    		}
    	}
    },
    "securityDefinitions": {  
    },
    "definitions": {
    	"Token":{
    		"type":"object",
    		"properties":{
    			"_id":{
                    "type": "string",
                    "example": "996b9a1a-c4c9-4b9a-ab3f-2b9bd8f89dc2"
    			},
    			"customerid":{
                    "type": "string",
                    "example": "uid0@email.com"
    			},
    			"lastAccessedTime":{
                    "type": "string",
                    "example": "2016-01-15T04:18:33.474Z"
    			},
    			"timeoutTime":{
                    "type": "string",
                    "example": "2016-01-16T04:18:33.474Z"
                }
    		}
    	}
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}